requirements.txt
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-community>=0.0.10
pydantic>=2.0.0
typer>=0.9.0
rich>=13.0.0
yfinance>=0.2.0
#duckduckgo-search>=4.0.0
cachetools>=5.0.0
python-dotenv>=1.0.0
ddgs